// BowlingBuddy/src/app/models/interfaces/ILeagueOverView.ts
export interface ILeagueOverView{
    ID: number;
    FileName: string;
    Name: string;
    Teams: number;
    Players: number;
    TotalCollected: number;
    LaneFeePerWeek: number;
    LaneFeesCollected: number;
    PrizeAmountPerWeek: number;
    PrizeAmountCollected: number;
    NumberOfWeeks: number;
    WeeksRecored: number;
}